% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createImputationData.R
\name{createImputationData}
\alias{createImputationData}
\title{Execute the Imputation study}
\usage{
createImputationData(outputFolder, TAR)
}
\arguments{
\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/)}

\item{TAR}{Time at risk for determining risk window}
}
\description{
Execute the Imputation study
}
\details{
This function will create the data for Imputation
}
